Label more Airflow metrics
Many Airflow metrics are ~ useless to users without labels to
group/filter them on. For example, duration metrics for deployments with
> 1 DAG commingle metrics for all DAGs, which makes it impossible to
spot regressions in the duration of a single DAG. Similarly, failure
counters with no labels prevent users from determining what failed. This
will increase the cardinality of the metrics we emit, but to do
otherwise would be to put the cart before the horse: some dimensions are
needed for the metrics to be useful.