Add tini to the alpine image
Having access to tini in a base image like this makes it convenient to
use the base image for utility containers that don't justify an image of
their own.

Specifically, a container that just runs a shell script with `/bin/sh -c
...` will usually have the problem that it doesn't forward SIGTERM (from
Kubernetes) to child processes.

With tini we can make such containers run `/sbin/tini -- /bin/sh -c ...`
and thus behave correctly in Kubernetes.